move config code
