+++ discovery_engine_core/core/src/lib.rs
@@ -30,13 +30,16 @@
     clippy::must_use_candidate
 )]
 
+mod config;
 pub mod document;
 mod engine;
 mod mab;
 mod ranker;
 pub mod stack;
-
 #[cfg(feature = "storage")]
 pub mod storage;
 
-pub use crate::engine::{Engine, Error, InitConfig, XaynAiEngine};
+pub use crate::{
+    config::InitConfig,
+    engine::{Engine, Error, XaynAiEngine},
+};
