+++ discovery_engine_core/core/src/stack/ops/personalized.rs
@@ -30,8 +30,8 @@ use xayn_discovery_engine_providers::{
 };
 
 use crate::{
+    config::EndpointConfig,
     document::{Document, HistoricDocument},
-    engine::EndpointConfig,
     stack::{
         filters::{ArticleFilter, CommonFilter, SourcesFilter},
         Id,
